now sorts dashboard articles by date (#69)
The df that we loop through doesn't have a 'date' column, the date gets instead added for each iteration in the loop. Div is also created for each loop, and this means that we needed to append the 'date'-object to the dictionary that holds each iteration of the loop, which later can be sorted outside of the main loop.